refactor: 패키지 구분
- model, view, util package 구분
