+++ src/main/java/calculator/Main.java
@@ -1,5 +1,7 @@
 package calculator;
 
+import calculator.model.CalculatorController;
+
 public class Main {
     public static void main(String[] args) {
         CalculatorController calculatorController = new CalculatorController();
