+++ src/main/java/calculator/exception/DividedNumberException.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.exception;
 
 public class DividedNumberException extends RuntimeException {
